move C/C++ interfaces
