+++ vst3cwrapper/src/main/public/cppinterface/IComponentHandler.h
@@ -1,24 +1,12 @@
 #pragma once
 
-#include "base/FUID.h"
-#include "base/PClassInfo.h"
-#include "base/PFactoryInfo.h"
+#include "FUnknown.h"
 #include "base/ftypes.h"
 #include "vst/IComponentHandler.h"
 
 #ifdef __cplusplus
 extern "C" {
 #endif
-
-typedef struct FUnknownVTable {
-    tresult PLUGIN_API (*queryInterface)(void *, const TUID, void **);
-    uint32 PLUGIN_API (*addRef)(void *);
-    uint32 PLUGIN_API (*release)(void *);
-} FUnknownVTable;
-
-typedef struct SFUnknown {
-    FUnknownVTable *vtable;
-} SFUnknown;
 
 typedef struct IComponentHandlerVTable {
     FUnknownVTable FUnknown;
@@ -32,9 +20,6 @@ typedef struct SIComponentHandler {
     IComponentHandlerVTable *vtable;
 } SIComponentHandler;
 
-SIComponentHandler *SIComponentHandler_alloc();
-void SIComponentHandler_free(SIComponentHandler *ptr);
-
 typedef struct IComponentHandler2VTable {
     FUnknownVTable FUnknown;
     tresult PLUGIN_API (*setDirty)(void *, TBool);
@@ -47,20 +32,11 @@ typedef struct SIComponentHandler2 {
     IComponentHandler2VTable *vtable;
 } SIComponentHandler2;
 
+// functions
+SIComponentHandler *SIComponentHandler_alloc();
 SIComponentHandler2 *SIComponentHandler2_alloc();
+void SIComponentHandler_free(SIComponentHandler *ptr);
 void SIComponentHandler2_free(SIComponentHandler2 *ptr);
-
-typedef struct IPluginFactoryVTable {
-    FUnknownVTable FUnknown;
-    tresult PLUGIN_API (*getFactoryInfo)(void *, PFactoryInfo *);
-    int32 PLUGIN_API (*countClasses)(void *);
-    tresult PLUGIN_API (*getClassInfo)(void *, int32, PClassInfo *);
-    tresult PLUGIN_API (*createInstance)(void *, FIDString, FIDString, void **);
-} IPluginFactoryVTable;
-
-typedef struct SIPluginFactory {
-    IPluginFactoryVTable *vtable;
-} SIPluginFactory;
 
 #ifdef __cplusplus
 }
